refactor: improve code structure and README
This commit improves the code structure by splitting the code over
multiple files. It also gives users the ability to use a local MongoDB
database during development.
