+++ .env.example
@@ -1,10 +1,12 @@
-MONGO_DB=
-MONGO_USERNAME=
-MONGO_PASSWORD=
-DISCORD_WEBHOOK_URL=
-API_TOKEN=
-TWITTER_CONSUMER_KEY=
-TWITTER_CONSUMER_SECRET=
-TWITTER_ACCESS_TOKEN_KEY=
-TWITTER_ACCESS_TOKEN_SECRET=
-INFURA_KEY=
+MONGO_USERNAME=                         # Database username
+MONGO_PASSWORD=                         # Database password
+MONGO_HOST=cluster0.eutpy.mongodb.net   # Database host
+MONGO_DB=                               # Database name
+DISCORD_WEBHOOK_URL=                    # Discord webhook url
+API_TOKEN=                              # Bearer Token for API
+INFURA_KEY=                             # Infura API Key
+NODE_ENV=                               # Environment (development, production)
+TWITTER_CONSUMER_KEY=                   # Twitter API Key
+TWITTER_CONSUMER_SECRET=                # Twitter API Secret
+TWITTER_ACCESS_TOKEN_KEY=               # Twitter Access Token
+TWITTER_ACCESS_TOKEN_SECRET=            # Twitter Access Token Secret
