+++ .gitignore
@@ -118,3 +118,7 @@ dist
 .yarn/build-state.yml
 .yarn/install-state.gz
 .pnp.*
+
+# IDE
+.vscode
+*.code-workspace
