+++ package.json
@@ -21,9 +21,10 @@
     "@types/mocha": "^9.1.1",
     "@vercel/node": "1.9.0",
     "chai": "^4.3.6",
-    "dotenv": "^16.0.0",
+    "dotenv": "^16.4.5",
     "mocha": "^9.2.2",
-    "ts-node": "^10.7.0"
+    "ts-node": "^10.7.0",
+    "vercel": "^23.1.2"
   },
   "keywords": [
     "serverless",
